facts/requirements.txt
chromadb
langchain
openai < 1
python-dotenv
tiktoken
